requirements.txt
# 3rd Party library
flake8
mypy
pydocstyle
pytest
pyinstaller
pywin32
openpyxl
pywebview
Dash
Plotly
pandas
numpy
bs4
GitPython
pysvn
mpi4py
opencv-python
PyQt6
PySide6
